Allow overriding start/stop characters for highlighting
Originally only Ms were highlighted, instead of highlighting things which are actually starts.
E.g. TTG, even though it is a recognised start codon in table 1, it translates to L by default.
The built in highlighter was not aware of the distinction and thus didn't highlight it as a possible
start codon.

Configuration options are supplied to allow users to override this setting, just like they can with .