Creating StaticMemoized store for sequences
This is intended for cases where the browser is expected not to cache information from the web server but still needs to cache this information specifically for performance reasons, for example where web services are used to serve sequence content.  Other than the caching behavior, it is a dropin replacement for Store/Sequence/StaticChunked.