Expose some params for user customisation.
Exposed 2 parameters to the user via the config file. They set the
minimum proportion ('indicatorProp') and minimum depth
('indicatorDepth') of alternative alleles required to render the
SNP indicator below a SNPCoverage track.

Default values are chosen so that indicators are displayed (almost)
as they would have been before this commit. The only difference is
that now SNP indicators will also display if the alternative alleles
are exactly half the total coverage at a site. This is due to a change
in a conditional from ">" to ">=".